chore(server): move seeder interface
